% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDragons.R
\name{collapseGenotype}
\alias{collapseGenotype}
\title{Sort alleles, so that aA or Aa is presented as Aa}
\usage{
collapseGenotype(dragon)
}
\description{
Sort alleles, so that aA or Aa is presented as Aa
}
