% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDragons.R
\docType{data}
\name{gene_names}
\alias{gene_names}
\title{Learn Mendelian inheritance with dragons and potatoes}
\format{
An object of class \code{character} of length 7.
}
\usage{
gene_names
}
\description{
Learn Mendelian inheritance with dragons and potatoes
}
\keyword{datasets}
